pyproject.toml
[project]
name = "alma-linux-remote-plugin"
version = "0.2.0"
description = "Alma Skill - Minimal stateless Linux remote for AI agents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "paramiko>=3.5.0",
    "pydantic>=2.9.0",
    "PyYAML>=6.0.1",
    "python-dotenv>=1.0.1"
]

[project.optional-dependencies]
dev = [
    "ruff>=0.5.0",
    "pytest>=8.0.0",
    "pytest-mock>=3.14.0",
    "pytest-cov>=5.0.0"
]

[dependency-groups]
dev = [
    "ruff>=0.5.0",
    "pytest>=8.0.0",
    "pytest-mock>=3.14.0",
    "pytest-cov>=5.0.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "--cov=src/alma_linux_remote_plugin --cov-report=term-missing"
python_files = "test_*.py"
testpaths = ["tests"]